Refactor allocation handler
